slight rework
